reorganize QC funcs
